requirements.txt
alembic==1.7.4
click==8.0.3
colorama==0.4.4
flake8==4.0.1
Flask==2.0.2
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.2
importlib-metadata==4.8.1
importlib-resources==5.3.0
itsdangerous==2.0.1
Jinja2==3.0.2
Mako==1.1.5
MarkupSafe==2.0.1
marshmallow==3.14.0
mccabe==0.6.1
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.3.0
python-dotenv==0.19.1
six==1.16.0
SQLAlchemy==1.4.26
Werkzeug==2.0.2
WTForms==2.3.3
zipp==3.6.0
